requirements.txt
requests==2.31.0
python-pptx==0.6.21
beautifulsoup4==4.12.2
python-dotenv==1.0.0
newsapi-python==0.2.7
yagmail==0.15.293
schedule==1.2.0
wikipedia==1.4.0
feedparser==6.0.10
selenium==4.11.2
webdriver_manager==4.0.1
openai==1.3.0
pyjokes==0.6.0
reportlab==4.0.8
python-pptx[pdf]==0.6.21 